+++ etils/edc/dataclass_utils.py
@@ -25,7 +25,9 @@ from typing import Any, Callable, TypeVar
 
 from etils import epy
 from etils.edc import cast_utils
+from etils.edc import context
 from etils.edc import frozen_utils
+from etils.edc import helpers
 
 _Cls = Any
 _ClsT = TypeVar('_ClsT')
@@ -40,6 +42,7 @@ def dataclass(
     replace: bool = ...,  # pylint: disable=redefined-outer-name
     repr: bool = ...,  # pylint: disable=redefined-builtin
     auto_cast: bool = ...,
+    contextvars: bool = ...,
     allow_unfrozen: bool = ...,
 ) -> Callable[[_ClsT], _ClsT]:
   ...
@@ -53,6 +56,7 @@ def dataclass(
     replace: bool = ...,  # pylint: disable=redefined-outer-name
     repr: bool = ...,  # pylint: disable=redefined-builtin
     auto_cast: bool = ...,
+    contextvars: bool = ...,
     allow_unfrozen: bool = ...,
 ) -> _ClsT:
   ...
@@ -65,6 +69,7 @@ def dataclass(
     replace=True,  # pylint: disable=redefined-outer-name
     repr=True,  # pylint: disable=redefined-builtin
     auto_cast=True,
+    contextvars=True,
     allow_unfrozen=False,
 ):
   """Augment a dataclass with additional features.
@@ -145,14 +150,47 @@ def dataclass(
     a.y.x  # Work
     ```
 
+  `contextvars`: Fields annotated as `edc.ContextVar` are wrapped in
+  a `contextvars.ContextVar`. Afterward each thread / asyncio coroutine will
+  have its own version of the fields (similarly to `threading.local`).
+
+  The contextvars are lazily initialized at first usage.
+
+  Example:
+
+  ```python
+  @edc.dataclass
+  @dataclasses.dataclass
+  class Context:
+    thread_id: edc.ContextVar[int] = dataclasses.field(
+        default_factory=threading.get_native_id
+    )
+    stack: edc.ContextVar[list[str]] = dataclasses.field(default_factory=list)
+
+  # Global context object
+  context = Context(thread_id=0)
+
+  def worker():
+    # Inside each thread, the worker use its own context
+    assert context.thread_id != 0
+    context.stack.append(1)
+
+  with concurrent.futures.ThreadPoolExecutor(max_workers=5) as executor:
+    for _ in range(10):
+      executor.submit(worker)
+  ```
+
   Args:
     cls: The dataclass to decorate
     kw_only: If True, make the dataclass `__init__` keyword-only.
     replace: If `True`, add a `.replace(` alias of `dataclasses.replace`.
     repr: If `True`, the class `__repr__` will return a pretty-printed `str`
       (one attribute per line)
-    auto_cast: If `True`, field annotated as `x: edc.AutoCast[Cls]` will be
+    auto_cast: If `True`, fields annotated as `x: edc.AutoCast[Cls]` will be
       converted to `x: Cls = edc.field(validator=Cls)`.
+    contextvars: It `True`, fields annotated as `x: edc.AutoCast[T]` are
+      converted to `contextvars`. This allow to have a `threading.local`-like
+      API for contextvars.
     allow_unfrozen: If `True`, add `.frozen`, `.unfrozen` methods.
 
   Returns:
@@ -181,8 +219,24 @@ def dataclass(
   if allow_unfrozen:
     cls = frozen_utils.add_unfrozen(cls)
 
+  descriptor_fns = []
   if auto_cast:
-    cls = cast_utils.apply_auto_cast_to_field(cls)
+    descriptor_fns.append(
+        helpers.DescriptorInfo(
+            annotation=cast_utils.AutoCast,
+            descriptor_fn=cast_utils.make_auto_cast_descriptor,
+        )
+    )
+
+  if contextvars:
+    descriptor_fns.append(
+        helpers.DescriptorInfo(
+            annotation=context.ContextVar,
+            descriptor_fn=context.make_contextvar_descriptor,
+        )
+    )
+
+  cls = helpers.wrap_new(cls, descriptor_fns)
 
   return cls